requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0

# Async HTTP
aiohttp>=3.8.0

# Progress bars
tqdm>=4.62.0

# For Jupyter/Colab compatibility
nest_asyncio>=1.5.0

# Google APIs (optional, for Drive integration)
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0

# Development
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
